backend/requirements.txt
# FastAPI and Web Framework
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
python-multipart>=0.0.5

# LLM Integration
litellm

# Document Processing
pillow
pymupdf

# Security and Validation
python-magic>=0.4.27
pydantic>=2.0.0
python-jose[cryptography]>=3.3.0

# Performance and Optimization
uvloop>=0.17.0  # High-performance event loop
aiofiles>=23.0.0  # Async file operations
redis>=4.5.0  # For distributed caching (optional)

# Monitoring and Logging
prometheus-client>=0.16.0
python-json-logger>=2.0.7

# Development Tools (optional)
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
ruff>=0.0.261